fix: move HashtagRepository to Repositories folder
